feat: restrict persistence ops to admin

+++ app/Http/Controllers/StudentController.php
@@ -7,6 +7,7 @@ use App\Exceptions\InvalidEmailException;
 use App\Exceptions\InvalidGenderException;
 use App\Exceptions\StudentNotFoundException;
 use App\Helpers\GlobalExceptionHandler;
+use App\Helpers\Utilities;
 use App\Services\Impl\StudentServiceImpl;
 use Illuminate\Http\Request;
 
@@ -26,6 +27,11 @@ class StudentController extends Controller
 
     public function store(Request $request)
     {
+
+        if (!Utilities::isAuthUserAdmin()) {
+            return response(['message' => 'Unauthorized'], 401);
+        }
+
         try
         {
             return response($this->studentService->createStudent($request->all()), 201);
@@ -50,6 +56,11 @@ class StudentController extends Controller
 
     public function update(int $id, Request $request)
     {
+
+        if (!Utilities::isAuthUserAdmin()) {
+            return response(['message' => 'Unauthorized'], 401);
+        }
+
         try
         {
             return response($this->studentService->updateStudentById($id, $request->all()), 200);
@@ -62,6 +73,11 @@ class StudentController extends Controller
 
     public function destroy(int $id)
     {
+
+        if (!Utilities::isAuthUserAdmin()) {
+            return response(['message' => 'Unauthorized'], 401);
+        }
+
         try
         {
             $this->studentService->deleteStudentById($id);
